Can we interrupt R to ask something to the user though the GUI?

Determine if R code execution can be interrupted by the GUI, e.g., using a modal dialog box. It depends both on R being in interactive() mode and the ask flag of the GUI being set to TRUE.

dont_ask(gui = .GUI)

dontAsk(gui = .GUI)

Arguments

gui
A gui object.

Value

TRUE if the GUI cannot interrupt R. The function triggering the dialog box should then not be displayed and it should return the default value as the result. The function returns TRUE if R is run in a non interactive session, or if ask is set to FALSE for the GUI, or if it is not specified (ask is NULL) then getOptions("gui.ask") is used.

Details

dontAsk and dont_ask are aliases.

See also

Examples

# What is the current state for the default GUI?
dont_ask()

#> [1] TRUE